# contracts/iot-ingestion.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus — IoT Ingestion
  version: 1.0.0
  description: |
    Contract for the IoT ingestion service used in FIT4110 Lab 03.
    This file is the source of truth for the Postman mock server and the Newman test collection.
  contact:
    name: Smart Campus Team
    email: lab@smartcampus.local
servers:
  - url: http://localhost:4010
    description: Local Prism mock server
  - url: https://api.campus.local
    description: Production placeholder
tags:
  - name: health
    description: Health check endpoints
  - name: readings
    description: IoT reading ingestion endpoints
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Service health check
      description: Returns a lightweight health payload so consumers can verify mock/server availability.
      tags: [health]
      security: []
      responses:
        '200':
          description: Service is running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              examples:
                ok:
                  value:
                    status: ok
                    service: iot-ingestion
                    time: '2026-05-28T08:00:00Z'

  /readings:
    post:
      operationId: createReading
      summary: Create a new sensor reading
      description: Accepts a new IoT device reading and returns a confirmation payload.
      tags: [readings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReadingRequest'
            examples:
              temperatureReading:
                value:
                  device_id: ESP32-LAB-A01
                  metric: temperature
                  value: 31.5
                  unit: celsius
                  timestamp: '2026-05-13T08:30:00+07:00'
      responses:
        '201':
          description: Reading created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingCreated'
              examples:
                created:
                  value:
                    reading_id: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                    device_id: ESP32-LAB-A01
                    metric: temperature
                    unit: celsius
                    value: 31.5
                    timestamp: '2026-05-13T08:30:00+07:00'
                    accepted: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /readings/latest:
    get:
      operationId: getLatestReadings
      summary: List latest readings
      description: Returns the most recent readings, optionally filtered by device_id and limited in size.
      tags: [readings]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Latest readings page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingPage'
              examples:
                latestReadings:
                  value:
                    items:
                      - reading_id: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                        device_id: ESP32-LAB-A01
                        metric: temperature
                        unit: celsius
                        value: 31.5
                        timestamp: '2026-05-13T08:30:00+07:00'
                        accepted: true
                    hasMore: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      required: [status, service, time]
      properties:
        status:
          type: string
          enum: [ok]
          example: ok
        service:
          type: string
          example: iot-ingestion
        time:
          type: string
          format: date-time
          example: '2026-05-28T08:00:00Z'

    CreateReadingRequest:
      type: object
      required: [device_id, metric, value, unit, timestamp]
      properties:
        device_id:
          type: string
          example: ESP32-LAB-A01
        metric:
          type: string
          enum: [temperature, humidity, pressure]
          example: temperature
        value:
          type: number
          minimum: -40
          maximum: 80
          example: 31.5
        unit:
          type: string
          enum: [celsius, percent, pascal]
          example: celsius
        timestamp:
          type: string
          format: date-time
          example: '2026-05-13T08:30:00+07:00'

    ReadingCreated:
      type: object
      required: [reading_id, device_id, metric, unit, value, timestamp, accepted]
      properties:
        reading_id:
          type: string
          format: uuid
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
        device_id:
          type: string
          example: ESP32-LAB-A01
        metric:
          type: string
          enum: [temperature, humidity, pressure]
          example: temperature
        unit:
          type: string
          enum: [celsius, percent, pascal]
          example: celsius
        value:
          type: number
          example: 31.5
        timestamp:
          type: string
          format: date-time
          example: '2026-05-13T08:30:00+07:00'
        accepted:
          type: boolean
          example: true

    ReadingPage:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReadingCreated'
        nextCursor:
          type: string
        hasMore:
          type: boolean

    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:                                                                
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string

  parameters:
    DeviceId:
      name: device_id
      in: query
      required: false
      schema:
        type: string
      description: Filter readings by device id.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of readings to return.

  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            missingField:
              value:
                type: https://smart-campus.local/problems/bad-request
                title: Bad Request
                status: 400
                detail: Thiếu trường bắt buộc
                instance: /readings
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            invalidToken:
              value:
                type: https://smart-campus.local/problems/unauthorized
                title: Unauthorized
                status: 401
                detail: Missing or invalid bearer token
                instance: /readings
    UnprocessableEntity:
      description: Request payload schema violation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            invalidMetric:
              value:
                type: https://smart-campus.local/problems/unprocessable-entity
                title: Unprocessable Entity
                status: 422
                detail: metric must be one of [temperature, humidity, pressure]
                instance: /readings
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            internalError:
              value:
                type: https://smart-campus.local/problems/internal-server-error
                title: Internal Server Error
                status: 500
                detail: Unexpected error
                instance: /readings